+++ previous-versions/commerce/src/codatcommerce/models/shared/__init__.py
@@ -2,15 +2,16 @@
 
 from .accountbalance import *
 from .address import *
-from .addresstype import *
+from .commerceaddresstype import *
+from .commercecustomerref import *
 from .companyinfo import *
 from .customer import *
-from .customerref import *
 from .customers import *
 from .dispute import *
 from .disputes import *
 from .disputestatus import *
-from .href import *
+from .errormessage import *
+from .halref import *
 from .links import *
 from .location import *
 from .locationref import *
@@ -21,7 +22,6 @@ from .orderlineitem import *
 from .orders import *
 from .payment import *
 from .paymentmethod import *
-from .paymentmethodref import *
 from .paymentmethods import *
 from .paymentmethodstatus import *
 from .paymentref import *
@@ -42,7 +42,6 @@ from .productvariant import *
 from .productvariantref import *
 from .productvariantstatus import *
 from .recordref import *
-from .schema import *
 from .security import *
 from .servicecharge import *
 from .servicechargetype import *
@@ -56,4 +55,4 @@ from .transactionsourceref import *
 from .transactionsourcetype import *
 from .transactiontype import *
 
-__all__ = ["AccountBalance","Address","AddressType","CompanyInfo","CompanyInfoWeblink","CompanyInfoWeblinkType","Customer","CustomerRef","Customers","Dispute","DisputeStatus","Disputes","HRef","Links","Location","LocationRef","Locations","Order","OrderDiscountAllocation","OrderLineItem","OrderSupplementalData","Orders","Payment","PaymentMethod","PaymentMethodRef","PaymentMethodStatus","PaymentMethods","PaymentRef","PaymentStatus","PaymentType","Payments","PhoneNumber","PhoneNumberType","Product","ProductCategories","ProductCategory","ProductInventory","ProductInventoryLocation","ProductPrice","ProductRef","ProductVariant","ProductVariantRef","ProductVariantStatus","Products","RecordRef","Schema","Security","ServiceCharge","ServiceChargeType","TaxComponent","TaxComponentAllocation","TaxComponentRef","TaxComponents","Transaction","TransactionSourceRef","TransactionSourceType","TransactionType","Transactions"]
+__all__ = ["AccountBalance","Address","CommerceAddressType","CommerceCustomerRef","CompanyInfo","CompanyInfoWeblink","CompanyInfoWeblinkType","Customer","Customers","Dispute","DisputeStatus","Disputes","ErrorMessage","HalRef","Links","Location","LocationRef","Locations","Order","OrderDiscountAllocation","OrderLineItem","OrderSupplementalData","Orders","Payment","PaymentMethod","PaymentMethodStatus","PaymentMethods","PaymentRef","PaymentStatus","PaymentType","Payments","PhoneNumber","PhoneNumberType","Product","ProductCategories","ProductCategory","ProductInventory","ProductInventoryLocation","ProductPrice","ProductRef","ProductVariant","ProductVariantRef","ProductVariantStatus","Products","RecordRef","Security","ServiceCharge","ServiceChargeType","TaxComponent","TaxComponentAllocation","TaxComponentRef","TaxComponents","Transaction","TransactionSourceRef","TransactionSourceType","TransactionType","Transactions"]