+++ previous-versions/commerce/src/codatcommerce/models/shared/taxcomponent.py
@@ -12,7 +12,7 @@ from typing import Optional
 @dataclasses.dataclass
 class TaxComponent:
     r"""The Tax Components endpoints return tax rates data from the commerce platform, including tax rate names and values. This is to support the mapping of tax rates from the commerce platform to those in the accounting platform.
-    
+
     Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=commerce-disputes) for this data type.
     """
     id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id') }})