+++ previous-versions/commerce/src/codatcommerce/models/shared/transaction.py
@@ -14,34 +14,34 @@ from typing import Optional
 @dataclasses.dataclass
 class Transaction:
     r"""Details of all financial transactions recorded in the commerce or point of sale system are added to the Transactions data type. For example, payments, service charges, and fees.
-    
+
     You can use data from the Transactions endpoints to calculate key metrics, such as:  
     - Transaction volumes  
     - Average transaction volume  
     - Average transaction value  
     - Returns  
     - Payouts
-    
+
     Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=commerce-transactions) for this data type.
     """
     id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id') }})
     r"""A unique, persistent identifier for this record"""
     created_date: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('createdDate'), 'exclude': lambda f: f is None }})
     r"""In Codat's data model, dates and times are represented using the <a class=\\"external\\" href=\\"https://en.wikipedia.org/wiki/ISO_8601\\" target=\\"_blank\\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
-    
+
     ```
     2020-10-08T22:40:50Z
     2021-01-01T00:00:00
     ```
-    
-    
-    
+
+
+
     When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
-    
+
     - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
     - Unqualified local time: `2021-11-15T01:00:00`
     - UTC time offsets: `2021-11-15T01:00:00-05:00`
-    
+
     > Time zones
     > 
     > Not all dates from Codat will contain information about time zones.  
@@ -49,30 +49,30 @@ class Transaction:
     """
     currency: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('currency'), 'exclude': lambda f: f is None }})
     r"""The currency data type in Codat is the [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217) currency code, e.g. _GBP_.
-    
+
     ## Unknown currencies
-    
+
     In line with the ISO 4217 specification, the code _XXX_ is used when the data source does not return a currency for a transaction. 
-    
+
     There are only a very small number of edge cases where this currency code is returned by the Codat system.
     """
     modified_date: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('modifiedDate'), 'exclude': lambda f: f is None }})
     source_created_date: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('sourceCreatedDate'), 'exclude': lambda f: f is None }})
     r"""In Codat's data model, dates and times are represented using the <a class=\\"external\\" href=\\"https://en.wikipedia.org/wiki/ISO_8601\\" target=\\"_blank\\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
-    
+
     ```
     2020-10-08T22:40:50Z
     2021-01-01T00:00:00
     ```
-    
-    
-    
+
+
+
     When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
-    
+
     - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
     - Unqualified local time: `2021-11-15T01:00:00`
     - UTC time offsets: `2021-11-15T01:00:00-05:00`
-    
+
     > Time zones
     > 
     > Not all dates from Codat will contain information about time zones.