+++ previous-versions/commerce/RELEASES.md
@@ -607,3 +607,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-commerce/0.33.1 - commerce
+
+## 2023-09-01 16:59:11
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/commerce
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-commerce/0.1.0 - previous-versions/commerce